# api/openapi/swagger.yaml
basePath: /
definitions:
...
  DataResponse:
    properties:
      data: {}
    type: object
...
  FestivalResponse:
    description: A named Hindu festival with its calendar date.
    properties:
      date:
        type: integer
      description:
        type: string
      festival:
        type: string
      id:
        type: string
    type: object
...
  ProblemDetail:
    properties:
      detail:
        type: string
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
info:
  contact:
    email: support@medha.app
    name: Medha Backend Team
    url: https://medha.app
  description: Backend API for the Medha platform. V2 Authentication (OTP-based),
    User Profiles, Events, Matching, and more.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: Medha API V2
  version: 2.0.0
paths:
...
  /api/v2/admin/festivals:
    get:
      description: Returns all festivals from the database.
      produces:
      - application/json
      responses:
        "200":
          description: List of festivals with id, festival, date
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data:
                  items:
                    additionalProperties: true
                    type: object
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - AdminToken: []
      summary: List festivals
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Manually create a new festival record.
      parameters:
      - description: Festival data
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            allOf:
            - {}
            - properties:
                date:
                  format: int64
                  type: integer
                festival:
                  type: string
              type: object
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Festival created
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data:
                  properties:
                    id:
                      type: string
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - AdminToken: []
      summary: Create festival
      tags:
      - admin
  /api/v2/admin/festivals/{id}:
    delete:
      description: Deletes a festival record. Returns 404 if festival not found.
      parameters:
      - description: Festival ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Festival deleted
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
        "404":
          description: Festival not found
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - AdminToken: []
      summary: Delete festival
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Update festival fields. Returns 404 if festival not found.
      parameters:
      - description: Festival ID
        in: path
        name: id
        required: true
        type: string
      - description: Festival data
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            allOf:
            - {}
            - properties:
                date:
                  format: int64
                  type: integer
                festival:
                  type: string
              type: object
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Festival updated
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data:
                  properties:
                    id:
                      type: string
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
        "404":
          description: Festival not found
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - AdminToken: []
      summary: Update festival
      tags:
      - admin
...
  /api/v2/panchanga/festivals:
    get:
      description: Returns upcoming Hindu festivals from today, ordered by date. Use
        `limit` to control results (default 100, max 500). No authentication required.
      parameters:
      - description: Max number of festivals to return (default 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Festival list
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/FestivalResponse'
                  type: array
              type: object
      summary: Upcoming Festivals
      tags:
      - panchanga-v2
...
securityDefinitions:
  AdminToken:
    description: Type "Bearer" followed by a space and an admin JWT or break-glass
      token.
    in: header
    name: Authorization
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...